libft

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: smihata <smihata@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 15:46:39 by smihata           #+#    #+#             */
-/*   Updated: 2023/03/02 16:03:07 by smihata          ###   ########.fr       */
+/*   Updated: 2023/03/08 16:43:25 by smihata          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 char	*ft_strdup(const char *str)
 {
 	char	*cpy;
-	int		len;
-	
+	size_t	len;
+
 	len = ft_strlen(str);
 	cpy = (char *)malloc(sizeof(char) * (len + 1));
 	if (!cpy)
